requirements.txt
edge_tts==7.0.2
langdetect==1.0.9
openai_whisper==20250625
protobuf==6.31.1
Requests==2.32.4
streamlit==1.45.1
streamlit_lottie==0.0.5
